+++ variables.tf
@@ -52,28 +52,31 @@ variable "subnet_ids" {
 }
 
 // http://docs.aws.amazon.com/AmazonECS/latest/developerguide/ecs-optimized_AMI_launch_latest.html
-variable "ecs_optimized_amis" {
-  description = "List with ecs optimized images per region, last updated on: 2018-12-17 (2018.03.i)."
-  type        = "map"
 
-  default = {
-    us-east-1      = "ami-045f1b3f87ed83659" # US East (N. Virginia)
-    us-east-2      = "ami-0307f7ccf6ea35750" # US East (Ohio)
-    us-west-1      = "ami-0285183bbef6224bd" # US West (N. California)
-    us-west-2      = "ami-01b70aea4161476b7" # US West (Oregon)
-    ca-central-1   = "ami-0f552e0a86f08b660" # Canada (Central)
-    eu-west-1      = "ami-0627e141ce928067c" # EU (Ireland)
-    eu-west-2      = "ami-01bee3897bba49d78" # EU (London)
-    eu-west-3      = "ami-0f4738fbeb53e6c3a" # EU (Paris)
-    eu-central-1   = "ami-0eaa3baf6969912ba" # EU (Frankfurt)
-    ap-northeast-1 = "ami-05b296a384694dfa4" # Asia Pacific (Tokyo)
-    ap-northeast-2 = "ami-00294948a592fc052" # Asia Pacific (Seoul)
-    ap-southeast-1 = "ami-050865a806e0dae53" # Asia Pacific (Singapore)
-    ap-southeast-2 = "ami-02c73ee1100ce3e7a" # Asia Pacific (Sydney)
-    ap-south-1     = "ami-01ef9f6a829ae3956" # Asia Pacific (Mumbai)
-    sa-east-1      = "ami-084b1eee100c102ee" # South America (São Paulo)
-    us-gov-west-1  = "ami-1dafcb7c"          # GovCloud
-  }
+variable "ecs_optimized_type" {
+  description = "Possible values"
+  default     = "amzn2"
+}
+
+variable "ecs_ami_filter" {
+  description = "The filter used to select the AMI for the ECS cluster. By default the the pattern `amzn2-ami-ecs-hvm-2.0.????????-x86_64-ebs` for the name is used."
+  type        = "list"
+
+  default = [{
+    name   = "name"
+    values = ["amzn2-ami-ecs-hvm-2.0.????????-x86_64-ebs"]
+  }]
+}
+
+variable "ecs_ami_latest" {
+  description = "Indicator to use the latest avaiable in the the list of the AMI's for the ECS cluster."
+  default     = true
+}
+
+variable "ecs_ami_owners" {
+  description = "A list of owners used to select the AMI for the ECS cluster."
+  type        = "list"
+  default     = ["amazon"]
 }
 
 variable "project" {